Updated Menu Calibration module and script
Updated the calibration module to include more subroutines:
initiateISUBNO, calculateEOF, validateEOF, initiateVarType,
initiateVarContainer, initiateVarLine and printResults. This will
essentially help organize the variables in case it needs to be expanded
in the future. The script was updated to incorporate the changes with
utilizing subroutines instead of doing all the pre-processing work
within the main program. Revision started on August 4; completed by
August 5, 2015.